// src/app/pages/simulaciones/models/parameter.ts
export class parameters {
    tituloReporte: string;
    fechaGeneracion: string;
    tipoPlan: Number;
    identificacionAsociado: string;
    nombreAsociado: string;
    nivelRiesgo: string;
    estado: string;
    edadIngreso: string;
    edadActual: string;
    fechaIngreso: string;
    fechaNacimiento: string;
    regional: string;
    oficinaVinculacion: string;
    nombrePlan: string;
    valorPlan: string;
    proteccionVida: string;
    valorProteccionVida: string;
    totalAportadoVida: string;
    valorTotalAportadoVida: string;
    reservaMatematica: string;
    valorReservaMate: string;
    anualidadVida: string;
    valorAnualidadVida: string;
    rescateVida: string;
    valorRescateVida: string;
    proteccionActualPerse: string;
    valorproteccionActualPerse: string;
    totalAportadoPerse: string;
    valorTotalAportadoPerse: string;
    textoValorAportPersevMuerte: string;
    valorAportPersevMuerte: string;
    textoPorcentajeAdm: string;
    porcentajeAdm: string;
    numeroCuota: string;
    numCuoPag: string;
    textoEdadSolicitud: string;
    edadSolicitud: string;
    textoPlanSolvencia: string;
    planSolvencia: string;
    textoporcentajeDevolucion: string;
    porcentajeDevolucion: string;
    procentajeRentabilidad: string;
    valorProcentajeRentabilidad: string;
    porcentajePerseverancia: string;
    valorPorcentajePerseverancia: string;
    rescatePerseverancia: string;
    valorRescatePerseverancia: string;
}